# backend/services/api-gateway/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PatientRecords API (stub)",
    "version": "0.1.0",
    "description": "Minimal API spec for development â€” includes auth and patient endpoints."
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Local dev" }
  ],
...
  "paths": {
...
    "/api/registry": {
      "get": {
        "tags": ["Registry"],
        "summary": "Get Module Registry",
        "description": "Retrieves the registry of available micro-frontend modules. No authentication required. Used by the shell app to discover and load dynamic modules.",
        "responses": {
          "200": {
            "description": "Module registry retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "version": { "type": "string", "example": "1.0.0" },
                    "description": { "type": "string" },
                    "modules": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "example": "demographics" },
                          "name": { "type": "string", "example": "Demographics" },
                          "description": { "type": "string" },
                          "icon": { "type": "string" },
                          "path": { "type": "string", "example": "demographics" },
                          "enabled": { "type": "boolean" },
                          "framework": { "type": "string", "enum": ["angular", "react"] },
                          "roles": { "type": "array", "items": { "type": "string" } },
                          "order": { "type": "number" },
                          "version": { "type": "string" },
                          "remoteEntry": { "type": "string", "example": "http://localhost:4201/remoteEntry.js" },
                          "remoteName": { "type": "string", "example": "demographicsApp" },
                          "exposedModule": { "type": "string", "example": "./DemographicsModule" }
                        },
                        "required": ["id", "name", "path", "enabled", "framework", "roles", "remoteEntry", "remoteName", "exposedModule"]
                      }
                    }
                  },
                  "required": ["version", "modules"]
                }
              }
            }
          },
          "500": {
            "description": "Failed to load registry"
          }
        }
      }
    },
    "/api/admin/registry": {
      "get": {
        "tags": ["Registry"],
        "summary": "Get Full Registry (Admin)",
        "description": "Retrieves the complete registry with all details. Admin access required.",
        "responses": {
          "200": { "description": "Full registry retrieved successfully" },
          "401": { "description": "Unauthorized - no token provided" },
          "403": { "description": "Forbidden - admin access required" }
        },
        "security": [ { "bearerAuth": [] } ]
      }
    },
    "/api/admin/registry/modules": {
      "post": {
        "tags": ["Registry"],
        "summary": "Add Module to Registry (Admin)",
        "description": "Add a new module to the registry. Admin access required.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "string", "example": "demographics" },
                  "name": { "type": "string", "example": "Demographics" },
                  "description": { "type": "string" },
                  "icon": { "type": "string" },
                  "path": { "type": "string" },
                  "enabled": { "type": "boolean" },
                  "roles": { "type": "array", "items": { "type": "string" } },
                  "order": { "type": "number" }
                },
                "required": ["id", "name"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Module added successfully" },
          "400": { "description": "Bad request - missing required fields" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden - admin access required" },
          "409": { "description": "Conflict - module already exists" }
        },
        "security": [ { "bearerAuth": [] } ]
      }
    },
    "/api/admin/registry/modules/{id}": {
      "put": {
        "tags": ["Registry"],
        "summary": "Update Module (Admin)",
        "description": "Update a module's properties. Admin access required.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Module ID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "enabled": { "type": "boolean" },
                  "order": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Module updated successfully" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden - admin access required" },
          "404": { "description": "Module not found" }
        },
        "security": [ { "bearerAuth": [] } ]
      },
      "delete": {
        "tags": ["Registry"],
        "summary": "Delete Module (Admin)",
        "description": "Remove a module from the registry. Admin access required.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Module ID" }
        ],
        "responses": {
          "200": { "description": "Module deleted successfully" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden - admin access required" },
          "404": { "description": "Module not found" }
        },
        "security": [ { "bearerAuth": [] } ]
      }
    },
    "/api/admin/registry/modules/{id}/toggle": {
      "patch": {
        "tags": ["Registry"],
        "summary": "Toggle Module Enable/Disable (Admin)",
        "description": "Enable or disable a module. Admin access required. If 'enabled' not specified, toggles current state.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Module ID" }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enabled": { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Module status updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "module": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "name": { "type": "string" },
                        "enabled": { "type": "boolean" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden - admin access required" },
          "404": { "description": "Module not found" }
        },
        "security": [ { "bearerAuth": [] } ]
      }
    },
...
    "/api/patients/{id}/visits": {
      "get": {
        "tags": ["Visits"],
        "summary": "Get visits for a patient",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "responses": { "200": { "description": "Array of visits" }, "404": { "description": "Not found" } },
        "security": [ { "bearerAuth": [] } ]
      },
      "post": {
        "tags": ["Visits"],
        "summary": "Create new visit record",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": { "type": "string" },
                  "visitType": { "type": "string", "enum": ["hospital", "clinic", "office"] },
                  "reason": { "type": "string" },
                  "notes": { "type": "string" },
                  "provider_name": { "type": "string" },
                  "facility_name": { "type": "string" },
                  "discharge_status": { "type": "string" }
                },
                "required": ["date", "visitType"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Visit created" }, "400": { "description": "Bad request" }, "404": { "description": "Patient not found" } },
        "security": [ { "bearerAuth": [] } ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  }
}